+++ swap/src/bob.rs
@@ -24,7 +24,6 @@ mod message0;
 mod message1;
 mod message2;
 mod message3;
-mod negotiate;
 pub mod swap;
 
 pub type Swarm = libp2p::Swarm<Behaviour>;
