+++ swap/src/bob/swap.rs
@@ -1,5 +1,5 @@
 use crate::{
-    bob::{event_loop::EventLoopHandle, negotiate::negotiate},
+    bob::event_loop::EventLoopHandle,
     state,
     state::{Bob, Swap},
     storage::Database,
@@ -17,7 +17,7 @@ use std::{convert::TryFrom, fmt, sync::Arc};
 use tracing::info;
 use uuid::Uuid;
 use xmr_btc::{
-    bob::{self},
+    bob::{self, State2},
     Epoch,
 };
 
@@ -392,4 +392,43 @@ where
             BobState::XmrRedeemed => Ok(BobState::XmrRedeemed),
         }
     }
+}
+
+pub async fn negotiate<R>(
+    state0: xmr_btc::bob::State0,
+    amounts: SwapAmounts,
+    swarm: &mut EventLoopHandle,
+    addr: Multiaddr,
+    mut rng: R,
+    bitcoin_wallet: Arc<crate::bitcoin::Wallet>,
+) -> Result<(State2, PeerId)>
+where
+    R: RngCore + CryptoRng + Send,
+{
+    tracing::trace!("Starting negotiate");
+    swarm.dial_alice(addr).await?;
+
+    let alice_peer_id = swarm.recv_conn_established().await?;
+
+    swarm
+        .request_amounts(alice_peer_id.clone(), amounts.btc)
+        .await?;
+
+    swarm
+        .send_message0(alice_peer_id.clone(), state0.next_message(&mut rng))
+        .await?;
+    let msg0 = swarm.recv_message0().await?;
+    let state1 = state0.receive(bitcoin_wallet.as_ref(), msg0).await?;
+
+    swarm
+        .send_message1(alice_peer_id.clone(), state1.next_message())
+        .await?;
+    let msg1 = swarm.recv_message1().await?;
+    let state2 = state1.receive(msg1)?;
+
+    swarm
+        .send_message2(alice_peer_id.clone(), state2.next_message())
+        .await?;
+
+    Ok((state2, alice_peer_id))
 }
